requirements.txt
pdfplumber==0.5.27
python-pptx==0.6.18
python-docx==0.8.10
xlrd==2.0.1
